rename helpers to helper & do some modifications
