+++ contract/AElf.Contracts.MultiToken/TokenContract_Helper.cs
