+++ src/AElf.Kernel.Types/Block/Block.cs
@@ -36,10 +36,5 @@ namespace AElf.Kernel
         {
             return Header.GetHash();
         }
-
-        public byte[] GetHashBytes()
-        {
-            return Header.GetHashBytes();
-        }
     }
 }
