+++ src/AElf.Kernel.Types/Block/BlockBody.cs
@@ -7,21 +7,11 @@ namespace AElf.Kernel
     public partial class BlockBody : IBlockBody
     {
         public int TransactionsCount => Transactions.Count;
-        private Hash _blockBodyHash;
-
-        private Hash CalculateBodyHash()
-        {
-            // TODO: BlockHeader is useless.
-            if (BlockHeader == null)
-                throw new InvalidOperationException("Block header hash is null.");
-            _blockBodyHash = Hash.FromRawBytes(this.ToByteArray());
-            return _blockBodyHash;
-        }
 
         /// <inheritdoc/>
         public Hash GetHash()
         {
-            return _blockBodyHash ?? CalculateBodyHash();
+            return Hash.FromRawBytes(this.ToByteArray());
         }
     }
 }
