+++ src/AElf.Kernel.Types/Block/IBlock.cs
@@ -4,6 +4,5 @@
     {
         BlockHeader Header { get; }
         BlockBody Body { get; }
-        byte[] GetHashBytes();
     }
 }
