+++ src/AElf.Types/Helper/AddressHelper.cs
