+++ src/AElf.Types/Helper/ByteArrayHelper.cs
