+++ src/AElf.Types/Helper/ByteStringHelper.cs
