+++ test/AElf.Contracts.Vote.Tests/VoteContractTestHelper.cs
