+++ test/AElf.Sdk.CSharp.Tests.TestContract/TokenContract_Helper.cs
