+++ test/AElf.Types.Tests/Helper/ByteStringHelperTests.cs
@@ -2,7 +2,7 @@
 using Google.Protobuf;
 using Shouldly;
 
-namespace AElf.Types.Tests.Helpers
+namespace AElf.Types.Tests.Helper
 {
     public class ByteStringHelperTests
     {
